[fix/#86]: 마스킹 처리 버그 수정
