+++ src/main/java/com/wekids/backend/config/MaskingStrategyConfig.java
@@ -11,13 +11,13 @@ import org.springframework.context.annotation.Profile;
 public class MaskingStrategyConfig {
 
     @Bean
-    @Profile("release")
+    @Profile("test")
     public DataMaskingStrategy noMaskingStrategy() {
         return new NoMaskingStrategy();
     }
 
     @Bean
-    @Profile({"dev", "test"})
+    @Profile({"dev", "release"})
     public DataMaskingStrategy maskingStrategy() {
         return new MaskingStrategy();
     }
