+++ src/main/java/com/wekids/backend/member/dto/response/ParentResponse.java
@@ -4,7 +4,7 @@ import com.wekids.backend.account.domain.Account;
 import com.wekids.backend.design.domain.Design;
 import com.wekids.backend.design.domain.enums.CharacterType;
 import com.wekids.backend.design.domain.enums.ColorType;
-import com.wekids.backend.utils.masking.service.DataMaskingService;
+import com.wekids.backend.utils.masking.service.DataMaskingServiceImpl;
 import com.wekids.backend.member.domain.Parent;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -26,7 +26,7 @@ public class ParentResponse {
     CharacterType character;
     Long accountId;
 
-    public void applyMasking(DataMaskingService maskingService) {
+    public void applyMasking(DataMaskingServiceImpl maskingService) {
         this.name = maskingService.maskData(this.name);
         this.accountNumber = maskingService.maskData(this.accountNumber);
         this.balance = maskingService.maskBalance(new BigDecimal(this.balance));
