+++ src/main/java/com/wekids/backend/utils/masking/strategy/DataMaskingStrategy.java
@@ -1,5 +1,8 @@
 package com.wekids.backend.utils.masking.strategy;
 
+import java.math.BigDecimal;
+
 public interface DataMaskingStrategy {
     String mask(String data);
+    String maskBalance(BigDecimal balance);
 }
