+++ src/main/java/com/wekids/backend/utils/masking/strategy/MaskingStrategy.java
@@ -1,5 +1,7 @@
 package com.wekids.backend.utils.masking.strategy;
 
+import java.math.BigDecimal;
+
 public class MaskingStrategy implements DataMaskingStrategy {
     @Override
     public String mask(String data) {
@@ -14,5 +16,16 @@ public class MaskingStrategy implements DataMaskingStrategy {
         }
 
         return data.replaceAll(".(?=.{4})", "*");
+    }
+
+    @Override
+    public String maskBalance(BigDecimal balance) {
+        if (balance == null) return null;
+
+        String balanceStr = balance.toPlainString();
+        int length = balanceStr.length();
+
+        if (length <= 2) return "0";
+        return balanceStr.substring(0, 2) + "*".repeat(length - 2);
     }
 }
