+++ src/main/java/com/wekids/backend/utils/masking/strategy/NoMaskingStrategy.java
@@ -1,8 +1,15 @@
 package com.wekids.backend.utils.masking.strategy;
 
+import java.math.BigDecimal;
+
 public class NoMaskingStrategy implements DataMaskingStrategy {
     @Override
     public String mask(String data) {
         return data;
+    }
+
+    @Override
+    public String maskBalance(BigDecimal balance) {
+        return balance.toString();
     }
 }
